Rename mesh_visualizer to mesh.
