+++ tensorboard/plugins/mesh/metadata_test.py
@@ -21,8 +21,8 @@ from __future__ import print_function
 from mock import patch
 import six
 import tensorflow as tf
-from tensorboard.plugins.mesh_visualizer import metadata
-from tensorboard.plugins.mesh_visualizer import plugin_data_pb2
+from tensorboard.plugins.mesh import metadata
+from tensorboard.plugins.mesh import plugin_data_pb2
 from tensorboard.util import test_util
